refactor(contexts): separate hooks and types from provider files
To improve React Fast Refresh compatibility, this commit extracts custom hooks (e.g., `useUI`, `useMetadata`), context objects, and their associated TypeScript types from the React component files where their providers are defined.

- Hooks, types, and context objects are moved to their own files.
- Provider components remain in their files and now import the context.
- A central `filterIndex.ts` barrel file is used for the filter context.
- All components consuming these contexts have been updated to use the new modular import paths.
