+++ src/contexts/UIContext.tsx
@@ -43,15 +43,14 @@
  * @see {@link RecentlyEditedContext} for recently edited tracking
  */
 
-import React, {
-  createContext,
-  useContext,
-  useState,
-  useCallback,
-  useMemo,
-  useEffect,
-  ReactNode,
-} from 'react';
+import React, { useState, useCallback, useMemo, useEffect, ReactNode } from 'react';
+import {
+  UIContext,
+  type UIContextType,
+  type ViewType,
+  type LoadingStates,
+  type SkeletonConfig,
+} from '@/contexts/contexts-ui';
 import { createLogger } from '@/utils/logger';
 
 const logger = createLogger('UIContext');
@@ -59,98 +58,9 @@ const logger = createLogger('UIContext');
 /**
  * View type definition for dashboard display
  */
-export type ViewType = 'grid' | 'list';
+// Types moved to contexts-ui.ts for Fast Refresh optimization
 
-/**
- * Loading state types for different UI components
- */
-export interface LoadingStates {
-  /** Main content loading state */
-  isMainContentLoading: boolean;
-  /** Skeleton loading state for cards */
-  isSkeletonLoading: boolean;
-  /** Search results loading state */
-  isSearchLoading: boolean;
-  /** Filter options loading state */
-  isFilterOptionsLoading: boolean;
-  /** Page transition loading state */
-  isPageTransitionLoading: boolean;
-}
-
-/**
- * Skeleton configuration for different loading patterns
- */
-export interface SkeletonConfig {
-  /** Number of skeleton items to display */
-  count: number;
-  /** Animation speed (ms) */
-  animationSpeed: number;
-  /** Enable shimmer effect */
-  enableShimmer: boolean;
-  /** Skeleton height variant */
-  heightVariant: 'small' | 'medium' | 'large';
-}
-
-/**
- * Context interface for UI state management
- *
- * Provides comprehensive UI state management for the dashboard
- * with focus on visual presentation and loading states.
- *
- * @interface UIContextType
- * @since 2025-07-08
- */
-interface UIContextType {
-  /** Current view type (grid or list) */
-  viewType: ViewType;
-
-  /** Update view type with persistence */
-  setViewType: (type: ViewType) => void;
-
-  /** Loading states for various UI components */
-  loadingStates: LoadingStates;
-
-  /** Update individual loading state */
-  setLoadingState: (key: keyof LoadingStates, value: boolean) => void;
-
-  /** Update multiple loading states at once */
-  setLoadingStates: (states: Partial<LoadingStates>) => void;
-
-  /** Skeleton configuration */
-  skeletonConfig: SkeletonConfig;
-
-  /** Update skeleton configuration */
-  setSkeletonConfig: (config: Partial<SkeletonConfig>) => void;
-
-  /** Mobile detection state */
-  isMobileView: boolean;
-
-  /** Responsive breakpoint detection */
-  isTabletView: boolean;
-
-  /** Current screen width */
-  screenWidth: number;
-
-  /** Utility functions */
-  utils: {
-    /** Reset all loading states */
-    resetLoadingStates: () => void;
-
-    /** Enable skeleton loading with configuration */
-    enableSkeletonLoading: (config?: Partial<SkeletonConfig>) => void;
-
-    /** Disable skeleton loading */
-    disableSkeletonLoading: () => void;
-
-    /** Get optimal skeleton count for current view */
-    getOptimalSkeletonCount: () => number;
-
-    /** Check if any loading state is active */
-    isAnyLoading: () => boolean;
-  };
-}
-
-const UIContext = createContext<UIContextType | null>(null);
+// Interface and context moved to contexts-ui.ts for Fast Refresh optimization
 
 /**
  * Default loading states
@@ -384,10 +294,5 @@ export const UIProvider: React.FC<UIProviderProps> = ({ children }) => {
  * @returns UIContextType with UI state and functions
  * @throws Error if used outside of UIProvider
  */
-export const useUI = (): UIContextType => {
-  const context = useContext(UIContext);
-  if (!context) {
-    throw new Error('useUI must be used within a UIProvider');
-  }
-  return context;
-};
+// Hook moved to useUI.ts for React Fast Refresh optimization
+// Import with: import { useUI } from '@/contexts/useUI';
