+++ src/hooks/queries/useProjects.ts
@@ -9,7 +9,7 @@ import { useMemo, useEffect } from 'react';
 import { useQuery, useQueryClient } from '@tanstack/react-query';
 import { createLogger } from '@/utils/logger';
 import { DashboardValidSortField } from '@/features/dashboard/dashboard.constants';
-import { SortDirectionType } from '@/contexts/FilterProvider';
+import { SortDirectionType } from '@/contexts/filterIndex';
 import { StatusBreakdown } from '@/types/dashboard';
 import { queryKeys, ProjectQueryParams } from './queryKeys';
 import { useRenderGuard, useThrottledLogger } from '@/utils/renderGuards';
